analog comparator updates
Mostly for style and naming.
