+++ chips/sam4l/src/aes.rs
@@ -157,8 +157,6 @@ pub struct Aes<'a> {
     stop_index: Cell<usize>,
 }
 
-pub static mut AES: Aes<'static> = Aes::new();
-
 impl<'a> Aes<'a> {
     pub const fn new() -> Aes<'a> {
         Aes {
@@ -508,3 +506,5 @@ impl<'a> hil::symmetric_encryption::AES128CBC for Aes<'a> {
         self.set_mode(encrypting, ConfidentialityMode::CBC);
     }
 }
+
+pub static mut AES: Aes<'static> = Aes::new();
