+++ chips/sam4l/src/dac.rs
@@ -154,13 +154,6 @@ impl hil::dac::DacChannel for Dac {
             let mr = Mode::WORD::HalfWordTransfer + Mode::STARTUP.val(0xff) + Mode::CLKDIV.val(0x60)
                 + Mode::TRGEN::InternalTrigger + Mode::DACEN::SET;
             regs.mr.write(mr);
-
-            regs.mr.modify(Mode::TRGSEL.val(1));
-            regs.wpmr.modify(WriteProtectMode::WPEN.val(1));
-            let a = regs.mr.read(Mode::TRGSEL);
-            let b = regs.wpmr.read(WriteProtectMode::WPEN);
-            debug!("Modifying as a word {}", a);
-            debug!("Modifying as a value {}", b);
         }
         ReturnCode::SUCCESS
     }
