--- doc/syscalls/00007_acifc.md
@@ -1,54 +1,0 @@
----
-driver number: 0x00007
----
-
-# ACIFC (Analog Comparator Interface)
-
-## Overview
-
-The Analog Comparator Interface (ACIFC) controls a number of Analog Comparators (ACs) with identical behavior. Each Analog Comparator compares two voltages and gives an output depending on this comparison. ACs can be first of all be configured in the 'normal mode', in which each AC performs a comparison of two voltages. The other option for comparison is the 'window mode', in which a voltage can be compared against a window of two voltages.
-
-A specific AC is referred to as ACx, where x is any number from 0 to n, and n is the index of the last AC module. 
-
-## Command
-
-  * ### Command number: `0`
-
-    **Description**: Does the driver exist?
-
-    **Argument 1**: unused
-
-    **Argument 2**: unused
-
-    **Returns**: SUCCESS if it exists, otherwise ENODEVICE
-
-  * ### Command number: `1`
-
-    **Description**: Initialize the ACIFC by enabling the clock, activating the ACs (Analog Comparators). Currently in initialization always-on mode is enabled, allowing a measurement on an AC to be made quickly after a measurement is triggered, without waiting for the AC startup time. The drawback is that when the AC is always on the power dissipation will be higher.
-
-    **Argument 1**: unused
-
-    **Argument 2**: unused
-
-    **Returns**: SUCCESS if the initialization was succesful
-	
-
-  * ### Command number: `2`
-
-    **Description**: Do a comparison of two inputs, referred to as the positive input voltage and the negative input voltage (Vinp and Vinn). 
-
-    **Argument 1**: The index of the Analog Comparator for which the comparison needs to be made, starting at 0.
-
-    **Argument 2**: unused
-
-    **Returns**: The output of this function is 1 when Vinp > Vinn, and 0 if Vinp < Vinn.
-
-  * ### Command number: `3`
-
-    **Description**: Do a comparison of three input voltages. Two ACs, ACx and ACx+1 are grouped for this comparison. The sources of the negative input of ACx (Vn_x) and the positive input of ACx+1 (Vp_x+1) must be connected together externally. These form the common voltage Vcommon. The other two remaining sources, being the positive input of ACx (Vp_x) and negative input of ACx+1 (Vn_x+1) then define an upper and a lower bound of a window. The result then depends on Vcommon lying inside of outside of this window. 
-
-    **Argument 1**: The index of the window for which to do a window comparison, starting at 0. For example, window 0 is the combination of ACx and ACx+1, window 1 is the combination of ACx+2 and ACx+3 etcetera.
-
-    **Argument 2**: unused
-
-    **Returns**: When the voltage of Vcommon lies inside the window defined by the positive input of ACx and the negative input of ACx+1, the output will be 1; it will be 0 if it is outside of the window. Specifically, the output will be 1 when Vn_x+1 < Vcommon < Vp_x, and 0 if Vcommon < Vn_x+1 or Vcommon > Vp_x.
