+++ doc/syscalls/README.md
@@ -49,7 +49,7 @@ stabilized or not (a "✓" indicates stability) in the Tock 1.0 release.
 |   | 0x00004       | [GPIO](00004_gpio.md)       | Set and read GPIO pins                     |
 | ✓ | 0x00005       | [ADC](00005_adc.md)         | Sample analog-to-digital converter pins    |
 |   | 0x00006       | DAC                         | Digital to analog converter                |
-|   | 0x00007       | [ACIFC](00007_acifc.md)     | Analog Comparator Interface                |
+|   | 0x00007       | [AnalogComparator](00007_analog_comparator.md) | Analog Comparator       |
 
 ### Kernel
